+++ pytorch_segmentation_models_trainer/config.py
@@ -1,0 +1,7 @@
+from pydantic import BaseSettings
+
+
+class Settings(BaseSettings):
+    app_name: str = "Awesome API"
+    config_path: str
+    config_name: str